backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
aiosqlite==0.22.1
pydantic-settings==2.12.0
redis==7.1.0
rq==2.6.1
faster-whisper==1.2.1
requests>=2.31.0
yt-dlp==2026.1.29
